Hazard edits, many platforms added, and melee enemy added
there is now a kill box attached to the camera and the melle enemy doe not have funtionality yet

+++ One Hop Bob/Assets/Scripts/PlatformMove.cs
@@ -6,7 +6,7 @@ public class PlatformMove : MonoBehaviour {
 
 
     public GameObject platform;
-    public int Movespeed;
+    public float Movespeed;
     public Vector3 currentpos;
     public Transform[] points;
     bool touched;
